+++ commands/bootstrap/bootstrap.go
@@ -76,6 +76,9 @@ func init() {
 	bootstrapCmd.PersistentFlags().StringVarP(&workingDirectoryPath,
 		"working_directory", "w", "work", "Working directory where to place deployment files")
 	viper.BindPFlag("working_directory", bootstrapCmd.PersistentFlags().Lookup("working_directory"))
+	bootstrapCmd.PersistentFlags().BoolVarP(&insecure, "insecure", "", false,
+		"Insecure mode - no TLS configuration")
+	viper.BindPFlag("review", bootstrapCmd.PersistentFlags().Lookup("review"))
 	bootstrapCmd.PersistentFlags().StringVarP(&deploymentID,
 		"deployment_name", "n", "", "Name of the deployment. If not specified deployment name is based on time.")
 	viper.BindPFlag("deployment_name", bootstrapCmd.PersistentFlags().Lookup("deployment_name"))
@@ -126,6 +129,7 @@ var reviewInputs bool
 var resourcesZipFilePath string
 var workingDirectoryPath string
 var inputsPath string
+var insecure bool
 var deploymentID string
 var configOnly bool